Lifting state up to the Parent Component
